fast-api-image-classification/requirements.txt
fastapi==0.105.0
tensorflow-cpu==2.14.0
cloudpickle==2.2.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
boto3==1.24.28
gunicorn==21.2.0